web_pylint/src/requirements/dev.txt
django
djangorestframework
markdown
django-filter
uwsgi
MySQL-python
mkdocs
docopt
ipython
django-extensions